+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 ### :house: Internal
 - `codeBlock`
   - [#527](https://github.com/wix-incubator/rich-content/pull/527) remove onTab (moved to keyBindingFn/handleKeyCommand)
+- `pubsub`
+  - [#528](https://github.com/wix-incubator/rich-content/pull/528) rename visibleBlock to focusedBlock
   
 </details>
 <hr/>